feat(App.js): add support for signup page feat(App.js): refactor isLoginPage to isAuthPage and add support for /signup page The isLoginPage function has been refactored to isAuthPage to better reflect its purpose. Support for the /signup page has been added to the function. The Sidenav component is now only rendered when the user is not on an authentication page.
feat(Booking.js): add list of cities to select from
The demarkRegion constant has been replaced with a list of cities. The list is now used to populate the "From" and "To" select fields.

feat(Login.js): add link to signup page
The "Don't have an account? Sign Up" link now navigates to the /signup page.

feat(Signup.js): add Signup page component
This commit adds a new Signup page component that allows users to sign up for the application. The component includes a form with fields for first name, last name, email, password, and gender. The form also includes a submit button that sends the form data to the server and redirects the user to the home page. The component uses Material-UI for styling and includes an image of a cargo ship.